Roll the view by angle for incap players
This could be improved somewhat if the rolling was set by the player's last movement vector instead of their angle

In other words: Unrolled would be the player's crawling direction, and rolled would be perpendicular.

Unfortunately I'm not sure if HDPlayerPawn has a variable that tracks last movement vector smoothly, and I wanted to keep this change simple for now.